DATAMONGO-1165 - Add support for Streaming large result lists.
Added support for translation of MongoDB Exceptions as well as potential event propagation via ReadDbObjectCallback in MongoTemplate.executeAsStream(..).
