+++ spring-data-mongodb/src/main/java/org/springframework/data/mongodb/core/MongoTemplate.java
@@ -95,7 +95,6 @@ import org.springframework.data.mongodb.core.query.Criteria;
 import org.springframework.data.mongodb.core.query.NearQuery;
 import org.springframework.data.mongodb.core.query.Query;
 import org.springframework.data.mongodb.core.query.Update;
-import org.springframework.data.mongodb.util.CloseableIterableCusorAdapter;
 import org.springframework.data.util.CloseableIterator;
 import org.springframework.jca.cci.core.ConnectionCallback;
 import org.springframework.util.Assert;
@@ -105,6 +104,7 @@ import org.springframework.util.StringUtils;
 
 import com.mongodb.BasicDBObject;
 import com.mongodb.CommandResult;
+import com.mongodb.Cursor;
 import com.mongodb.DB;
 import com.mongodb.DBCollection;
 import com.mongodb.DBCursor;
@@ -318,7 +318,7 @@ public class MongoTemplate implements MongoOperations, ApplicationContextAware {
 	 * @see org.springframework.data.mongodb.core.MongoOperations#executeAsStream(org.springframework.data.mongodb.core.query.Query, java.lang.Class)
 	 */
 	@Override
-	public <T> CloseableIterator<T> executeAsStream(final Query query, final Class<?> entityType) {
+	public <T> CloseableIterator<T> executeAsStream(final Query query, final Class<T> entityType) {
 
 		return execute(entityType, new CollectionCallback<CloseableIterator<T>>() {
 
@@ -331,8 +331,9 @@ public class MongoTemplate implements MongoOperations, ApplicationContextAware {
 				DBObject mappedQuery = queryMapper.getMappedObject(query.getQueryObject(), persistentEntity);
 
 				DBCursor cursor = collection.find(mappedQuery, mappedFields);
+				ReadDbObjectCallback<T> readCallback = new ReadDbObjectCallback<T>(mongoConverter, entityType);
 
-				return new CloseableIterableCusorAdapter<T>(cursor, mongoConverter, exceptionTranslator, entityType);
+				return new CloseableIterableCusorAdapter<T>(cursor, exceptionTranslator, readCallback);
 			}
 		});
 
@@ -2123,9 +2124,10 @@ public class MongoTemplate implements MongoOperations, ApplicationContextAware {
 	 * Simple internal callback to allow operations on a {@link DBObject}.
 	 * 
 	 * @author Oliver Gierke
+	 * @author Thomas Darimont
 	 */
 
-	private interface DbObjectCallback<T> {
+	static interface DbObjectCallback<T> {
 
 		T doWith(DBObject object);
 	}
@@ -2285,6 +2287,79 @@ public class MongoTemplate implements MongoOperations, ApplicationContextAware {
 			T doWith = delegate.doWith(content);
 
 			return new GeoResult<T>(doWith, new Distance(distance, metric));
+		}
+	}
+
+	/**
+	 * A {@link CloseableIterator} that is backed by a MongoDB {@link Cursor}.
+	 * 
+	 * @since 1.7
+	 * @author Thomas Darimont
+	 */
+	static class CloseableIterableCusorAdapter<T> implements CloseableIterator<T> {
+
+		private volatile Cursor cursor;
+		private PersistenceExceptionTranslator exceptionTranslator;
+		private DbObjectCallback<T> objectReadCallback;
+
+		/**
+		 * Creates a new {@link CloseableIterableCusorAdapter} backed by the given {@link Cursor}.
+		 * 
+		 * @param cursor
+		 * @param exceptionTranslator
+		 * @param objectReadCallback
+		 */
+		public CloseableIterableCusorAdapter(Cursor cursor, PersistenceExceptionTranslator exceptionTranslator,
+				DbObjectCallback<T> objectReadCallback) {
+
+			this.cursor = cursor;
+			this.exceptionTranslator = exceptionTranslator;
+			this.objectReadCallback = objectReadCallback;
+		}
+
+		@Override
+		public boolean hasNext() {
+
+			if (cursor == null) {
+				return false;
+			}
+
+			try {
+				return cursor.hasNext();
+			} catch (RuntimeException ex) {
+				throw exceptionTranslator.translateExceptionIfPossible(ex);
+			}
+		}
+
+		@Override
+		public T next() {
+
+			if (cursor == null) {
+				return null;
+			}
+
+			try {
+				DBObject item = cursor.next();
+				T converted = objectReadCallback.doWith(item);
+				return converted;
+			} catch (RuntimeException ex) {
+				throw exceptionTranslator.translateExceptionIfPossible(ex);
+			}
+		}
+
+		@Override
+		public void close() {
+
+			Cursor c = cursor;
+			try {
+				c.close();
+			} catch (RuntimeException ex) {
+				throw exceptionTranslator.translateExceptionIfPossible(ex);
+			} finally {
+				cursor = null;
+				exceptionTranslator = null;
+				objectReadCallback = null;
+			}
 		}
 	}
 }
